Thesis_CrossLingual_GovQA/requirements.txt
pdfplumber>=0.9.0
requests>=2.31.0
openai>=1.0.0
google-generativeai>=0.8.0
mistralai>=1.0.0
groq>=0.4.0
bert-score>=0.3.13
sacrebleu>=2.3.1
rouge-score>=0.1.2
numpy>=1.24.0
pandas>=2.0.0
